+++ src/app/timelogs/datepicker/custom-header/custom-header.component.css
@@ -1,0 +1,19 @@
+.custom-header {
+  display: flex;
+  align-items: center;
+  padding: 0.5em;
+}
+.custom-header button {
+  margin: 5px;
+}
+.custom-header .custom-header-label {
+  flex: 1;
+  height: 1em;
+  font-weight: 500;
+  text-align: center;
+}
+.custom-header .custom-double-arrow .mat-icon {
+  margin: -22%;
+}
+
+/*# sourceMappingURL=custom-header.component.css.map */
