+++ src/app/timelogs/datepicker/custom-header/custom-header.component.html
@@ -1,0 +1,10 @@
+<div class="custom-header">
+  <button mat-stroked-button color="primary" (click)="todayClicked()">Today</button>
+  <span class="custom-header-label">{{periodLabel}}</span>
+  <button mat-icon-button (click)="previousClicked()">
+    <mat-icon>keyboard_arrow_left</mat-icon>
+  </button>
+  <button mat-icon-button (click)="nextClicked()">
+    <mat-icon>keyboard_arrow_right</mat-icon>
+  </button>
+</div>
