+++ src/app/timelogs/datepicker/custom-header/custom-header.component.sass
@@ -1,0 +1,16 @@
+.custom-header
+  display: flex
+  align-items: center
+  padding: 0.5em
+
+  button
+    margin: 5px
+
+  .custom-header-label
+    flex: 1
+    height: 1em
+    font-weight: 500
+    text-align: center
+
+  .custom-double-arrow .mat-icon
+    margin: -22%
